renamed because apparently "global" is something pre-defined in python
