+++ GANDLF/data/augmentation/all_defines.py
